pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gazzeton"
version = "0.1.0"
description = "Deterministic RSS feed aggregator with Markdown output"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0",
    "PyYAML>=6.0",
    "click>=8.0",
]

[project.scripts]
gazzeton = "gazzeton.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
